EXTKeyPathCoding: Change `@keypath` to `$keypath`.
Xcode 10 changed the return value of the box operator (`@`) to return
a `nullable NSString *` instead of `NSString *` when the parameter is
`char *`. This change makes the `@keypath` macro return nullable value
as well.

When enabling a compiler warning to warn whenever a nullable variable
pointer is set to a non-nullable variable, many warning are thrown due
to the above change.

One solution is to surround each `@keypath` usage with `nn()`, but
that would be very annoying. Since the warning's source is outside the
original macro, it cannot be fixed inside the macro while still using
`@`. The solution with the least amount of change is this.