EXTKeyPathCoding: decrease runtime cost of keypath macros.
The keypath macros `@keypath` and `@collectionKeypath` used to have a
non negligible runtime overhead, as they allocated Obj-C objects and
sent messages although they are only used to ensure compile-time safety.

For example, let's take a look at `@keypath` macro with two arguments
(the common case). The code `@keypath(foo, bar)` expands to:

```
@(NO).boolValue ? @"" :
  ((NSString * _Nonnull)@((((void)(NO && ((void)foo.bar, NO)), "bar"))))
```

Or, after compilation with `-Os`, the code looks like this:
```
NSString *result;
if ([[NSNumber numberWithBool:NO] boolValue]) {
  result = @"";
} else {
  result = [NSString stringWithUTF8String:"baz"];
}
```

This code has two allocations: one for `NSNumber` and one for
`NSString`. Moreover, there's work needs to be done to unbox the `BOOL`
value and to initialize the string (that probably require, for the very
least, to calculate its length).

Additionally, because of the dynamic nature of Obj-C, once there are any
dynamic dispatches, this code cannot be optimized by the compiler (even
if we know that `[[NSNumber numberWithBool:NO] boolValue]` is _always_
false).

To mitigate that, a new formulation of the macro that uses `@selector`
is proposed. The cost of creating a predefined selector is very cheap,
and involves only a single QWORD load (on 64-bit platforms), which is
the pointer to the interned `char *` that represents that selector.

The compiled code for the new macro looks like this:
```
NSString *result = @"baz";
if (!@selector(alloc)) {
  result = @"";
}
```

Which generates less code and is more space efficient and doesn't
contain any message sends, so the compiler can also optimize this away
if needed.

A microbenchmark shows a 90% run time reduction for the new
implementation of the `@keypath` macro.