EXTNil: remove from project.
This class contains a design issue that cannot be solved: when
allocating buffer used to return data to the caller, it tries to guess
the requested return size by iterating over all the loaded classes and
see whether they respond to the selector representing the message sent
to this object. If the object responds to that selector, the method
signature that deals with this message send is used in order to
determine the return size.

However, a single selector may back up different implementations, making
`EXTNil` return uninitialized memory. One example is the
`-rangeOfString:` selector that exists in the test suite.

Since we're not using this anywhere, it is better to just remove this
class.